requirements.txt
langchain==0.0.205
openai==0.27.6
faiss-cpu==1.7.4
streamlit==1.23.1
streamlit-chat==0.0.2.2
tiktoken==0.3.3

